Move name_check to its own mod
